+++ cmake/share/StandaloneC.cmake
@@ -31,7 +31,7 @@ if( STANDALONE)
    endif()
 
    if( NOT STANDALONE_DEFINITIONS)
-      set( STANDALONE_DEFINITIONS ${mulle_objc_compat_DEFINITIONS})
+      set( STANDALONE_DEFINITIONS ${MULLE_OBJC_COMPAT_DEFINITIONS})
    endif()
 
    #